Create FirstExample
This is realisation of creating basic neural network with 3 layers: One input layer, One hidden layer and One output layer. The main idea of network to get an answer on some question (in "True/False" format), which is based on factors, which literally presented in input layer. 

IMPORTANT: There is no learning proses in this network. Learning proses of neural network means selection of weight values ​​by a neural network (sometimes by 'random'). But in this case we give values of weights by ourselves.